feat: Update field types in attrs.py
This change updates the data types of several fields in the `FIELDS` dictionary in `attrs.py`:

1. Changes the data type of `frequencia_inicial` and `frequencia_final` fields from `SimpleField` to `SimpleField` with `_dtype="float"`.
2. Changes the data type of `horas_de_conclusao`, `horas_de_deslocamento`, `horas_de_execucao`, and `horas_de_preparacao` fields from `SimpleField` with `_dtype="int"` to `SimpleField` with `_dtype="float"`.
3. Changes the data type of `latitude_coordenadas`, `latitude_da_estacao`, `longitude_coordenadas`, and `longitude_da_estacao` fields from `SimpleField` to `SimpleField` with `_dtype="float"`.

These changes ensure that the fields are properly handled as floating-point values, improving the overall data integrity and consistency in the application.

+++ fiscaliza/attrs.py
@@ -87,8 +87,8 @@ FIELDS = {
     "foi_constatada_interferencia": FieldWithOptions(
         1967, "Foi constatada interferência?", mandatory=True, options=["", "0", "1"]
     ),
-    "frequencia_inicial": SimpleField(156, "Frequência inicial:", True),
-    "frequencia_final": SimpleField(158, "Frequência final:", True),
+    "frequencia_inicial": SimpleField(156, "Frequência inicial:", True, _dtype="float"),
+    "frequencia_final": SimpleField(158, "Frequência final:", True, _dtype="float"),
     "frequencias": SimpleField(180, "Frequência(s):"),
     "gerar_plai": FieldWithOptions(
         426,
@@ -102,12 +102,12 @@ FIELDS = {
         options=["", "0", "1"],
         mapping={"1": ["html"]},
     ),
-    "horas_de_conclusao": SimpleField(94, "Horas de conclusão", True, _dtype="int"),
+    "horas_de_conclusao": SimpleField(94, "Horas de conclusão", True, _dtype="float"),
     "horas_de_deslocamento": SimpleField(
-        92, "Horas de deslocamento", True, _dtype="int"
+        92, "Horas de deslocamento", True, _dtype="float"
     ),
-    "horas_de_execucao": SimpleField(93, "Horas de Execução", True, _dtype="int"),
-    "horas_de_preparacao": SimpleField(91, "Horas de Preparação", True, _dtype="int"),
+    "horas_de_execucao": SimpleField(93, "Horas de Execução", True, _dtype="float"),
+    "horas_de_preparacao": SimpleField(91, "Horas de Preparação", True, _dtype="float"),
     "houve_interferencia": FieldWithOptions(
         149,
         "Houve interferência?",
@@ -140,10 +140,14 @@ FIELDS = {
         format_value=True,
     ),
     "lai_vinculadas": SimpleField(481, "LAI vinculadas"),
-    "latitude_coordenadas": SimpleField(170, "Latitude (º):", True),
-    "latitude_da_estacao": SimpleField(191, "Latitude da estação:", True),
-    "longitude_coordenadas": SimpleField(171, "Longitude (º):", True),
-    "longitude_da_estacao": SimpleField(192, "Longitude da estação:", True),
+    "latitude_coordenadas": SimpleField(170, "Latitude (º):", True, _dtype="float"),
+    "latitude_da_estacao": SimpleField(
+        191, "Latitude da estação:", True, _dtype="float"
+    ),
+    "longitude_coordenadas": SimpleField(171, "Longitude (º):", True, _dtype="float"),
+    "longitude_da_estacao": SimpleField(
+        192, "Longitude da estação:", True, _dtype="float"
+    ),
     "motivo_de_lai": FieldWithOptions(
         164,
         "Motivo de LAI:",
